requirements.txt
Django==3.2.10
mysqlclient==2.1.0
djangorestframework==3.13.1
drf-yasg==1.20.0